pyproject.toml
[build-system]
requires = ["setuptools>=75.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "symbiote"
version = "0.1.0"
description = "Kernel for persistent cognitive entities"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"

dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0,<25.0.0",
    "httpx>=0.27.0,<0.28.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "pre-commit>=3.7.0",
]
llm = [
    "forge-llm @ git+https://github.com/symlabs-ai/forgellmclient.git",
]

[project.scripts]
symbiote = "symbiote.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["src/symbiote"]

[tool.coverage.report]
fail_under = 85
